pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-agent"
version = "0.1.0"
description = "Local-first stock trading agent with backtesting, paper trading, and optional live execution"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "Stock Agent Dev"}]

dependencies = [
    "pandas>=2.1.0",
    "numpy>=1.25.0",
    "scipy>=1.11.0",
    "statsmodels>=0.14.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.1.0",
    "matplotlib>=3.8.0",
    "plotly>=5.18.0",
    "duckdb>=0.9.0",
    "pyarrow>=14.0.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.5.0",
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
deep = ["torch>=2.1.0"]
dev = ["ruff>=0.1.0", "black>=23.0", "pytest>=7.4.0", "pre-commit>=3.5.0"]

[project.scripts]
sa-data = "src.cli.data_cli:app"
sa-backtest = "src.cli.backtest_cli:app"
sa-paper = "src.cli.paper_cli:app"
sa-live = "src.cli.live_cli:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
